enable gpu reader by default
