+++ src/op/scan/parquet_gpu_ingestible.cpp
@@ -73,24 +73,6 @@ struct rg_accumulator {
   std::optional<std::vector<std::string>> partition_values;
 };
 
-/// Strip a leading "file://" scheme (case-insensitive) so the path can be
-/// resolved by a local-file backend.
-std::string normalize_path(std::string const& p)
-{
-  static constexpr std::string_view kFile = "file://";
-  if (p.size() > kFile.size()) {
-    bool is_file_uri = true;
-    for (std::size_t i = 0; i < kFile.size(); ++i) {
-      if (std::tolower(static_cast<unsigned char>(p[i])) != static_cast<unsigned char>(kFile[i])) {
-        is_file_uri = false;
-        break;
-      }
-    }
-    if (is_file_uri) { return p.substr(kFile.size()); }
-  }
-  return p;
-}
-
 bool has_uri_scheme(std::string const& p) { return p.find("://") != std::string::npos; }
 
 }  // namespace
@@ -214,7 +196,7 @@ void parquet_gpu_ingestible::run_batch(file_batch const& batch,
         try {
           // Resolve the probe file to a datasource (carries its own io backend,
           // cache and metadata) instead of reaching into io_context.
-          auto probe_ds = _scan_manager->create_datasource(normalize_path(probe_path));
+          auto probe_ds = _scan_manager->create_datasource(probe_path);
           if (!probe_ds) { throw std::runtime_error("no backend supports path: " + probe_path); }
           std::shared_ptr<cudf::io::parquet::FileMetaData const> probe_meta;
           if (auto cached = probe_ds->metadata()) {
@@ -307,22 +289,14 @@ void parquet_gpu_ingestible::run_batch(file_batch const& batch,
       accum.partition_values = std::move(file_partition_values);
     }
 
-    auto const lookup_path = normalize_path(file_path);
     // Resolve the file to a sirius_datasource — it carries its own io backend,
     // prefetch cache and cached metadata, so the ingestible no longer reaches
     // into io_context. Fall back to a plain cudf datasource only for local
     // paths no sirius backend claims.
-    auto sirius_ds = _scan_manager->create_datasource(lookup_path);
-    if (!sirius_ds && has_uri_scheme(lookup_path)) {
+    auto sirius_ds = _scan_manager->create_datasource(file_path);
+    if (!sirius_ds && has_uri_scheme(file_path)) {
       throw std::runtime_error("[parquet_gpu_ingestible] no backend supports path: " + file_path);
     }
-    std::shared_ptr<sirius::io::sirius_io_object> file_io_object =
-      sirius_ds ? sirius_ds->io_object() : nullptr;
-    std::shared_ptr<sirius::io::sirius_ioctx> file_io_ctx =
-      sirius_ds ? sirius_ds->io_ctx() : nullptr;
-    std::shared_ptr<cudf::io::datasource> file_datasource =
-      sirius_ds ? std::static_pointer_cast<cudf::io::datasource>(sirius_ds)
-                : std::shared_ptr<cudf::io::datasource>(cudf::io::datasource::create(lookup_path));
 
     std::shared_ptr<cudf::io::parquet::FileMetaData const> file_metadata;
     std::shared_ptr<scan_manager::parquet_metadata> cached_parquet_metadata;
@@ -341,7 +315,7 @@ void parquet_gpu_ingestible::run_batch(file_batch const& batch,
       footer_byte_len = cached_parquet_metadata->footer_byte_len();
       reader_ptr      = std::make_unique<hybrid_scan_reader>(*file_metadata, *reader_options);
     } else {
-      auto footer_buffer = cudf::io::parquet::fetch_footer_to_host(*file_datasource);
+      auto footer_buffer = cudf::io::parquet::fetch_footer_to_host(*sirius_ds);
       footer_byte_len    = footer_buffer->size();
       reader_ptr         = std::make_unique<hybrid_scan_reader>(
         cudf::host_span<uint8_t const>(footer_buffer->data(), footer_buffer->size()),
@@ -387,61 +361,6 @@ void parquet_gpu_ingestible::run_batch(file_batch const& batch,
                        rgs_before - row_group_indices.size());
     }
 
-    // Prefetch cache prewarm — see parquet_split_provider.cpp:521-604.
-    if (file_io_object && file_io_ctx && file_io_ctx->cache() != nullptr &&
-        !row_group_indices.empty()) {
-      using range_t = cudf::io::text::byte_range_info;
-
-      auto chunk_ranges = reader.all_column_chunks_byte_ranges(row_group_indices, *reader_options);
-      std::sort(chunk_ranges.begin(), chunk_ranges.end(), [](auto const& a, auto const& b) {
-        return a.offset() < b.offset();
-      });
-      std::vector<range_t> merged;
-      merged.reserve(chunk_ranges.size());
-      if (!chunk_ranges.empty()) {
-        auto cur_start = chunk_ranges[0].offset();
-        auto cur_end   = cur_start + chunk_ranges[0].size();
-        for (auto const& r : chunk_ranges) {
-          auto const rs = r.offset();
-          auto const re = rs + r.size();
-          if (rs <= cur_end) {
-            cur_end = std::max(cur_end, re);
-          } else {
-            merged.emplace_back(cur_start, cur_end - cur_start);
-            cur_start = rs;
-            cur_end   = re;
-          }
-        }
-        merged.emplace_back(cur_start, cur_end - cur_start);
-      }
-
-      constexpr std::size_t FOOTER_TAIL_SIZE = 8;
-      auto const file_size                   = file_io_object->size();
-      auto const footer_off  = static_cast<int64_t>(file_size - FOOTER_TAIL_SIZE - footer_byte_len);
-      auto const footer_size = static_cast<int64_t>(FOOTER_TAIL_SIZE + footer_byte_len);
-
-      std::vector<range_t> ranges;
-      ranges.reserve(merged.size() + 2);
-      ranges.emplace_back(0, 4);
-      ranges.insert(ranges.end(), merged.begin(), merged.end());
-      ranges.emplace_back(footer_off, footer_size);
-      std::sort(ranges.begin(), ranges.end(), [](auto const& a, auto const& b) {
-        return a.offset() < b.offset();
-      });
-
-      std::shared_ptr<sirius::io::sirius_io_object_metadata> metadata_to_store =
-        cached_parquet_metadata
-          ? nullptr
-          : std::static_pointer_cast<sirius::io::sirius_io_object_metadata>(
-              std::make_shared<scan_manager::parquet_metadata>(file_metadata, footer_byte_len));
-      bool const prewarm = (_scan_manager != nullptr) && _scan_manager->chunk_prewarm_enabled();
-      if (prewarm) {
-        file_io_ctx->cache()->insert(*file_io_object, std::move(metadata_to_store), ranges);
-      } else {
-        file_io_ctx->cache()->insert(*file_io_object, std::move(metadata_to_store), {});
-      }
-    }
-
     std::vector<cudf::size_type> cur_rgs;
     std::size_t cur_uncompressed_bytes = 0;
     std::size_t cur_compressed_bytes   = 0;
@@ -453,7 +372,7 @@ void parquet_gpu_ingestible::run_batch(file_batch const& batch,
                                 std::move(cur_rgs),
                                 cur_uncompressed_bytes,
                                 cur_compressed_bytes,
-                                file_datasource);
+                                sirius_ds);
       accum.total_uncompressed_bytes += cur_uncompressed_bytes;
       cur_rgs.clear();
       cur_uncompressed_bytes = 0;
