+++ test/cpp/integration/test_gpu_execution_tpch.cpp
@@ -313,23 +313,24 @@ class GPUExecutionDuckDBFixture : public GPUExecutionFixtureBase {
     REQUIRE_FALSE(result->HasError());
   }
 
-  // Disabled: these tests scan native (DuckDB-storage) tables via seq_scan. The
-  // legacy duckdb_scan path was removed and GPU native scan has no IO backend
-  // wired in this harness ("missing io_ctx, io_obj"), so every query throws
-  // "Unsupported scan function: seq_scan" and poisons the shared integration
-  // DB. Shadow the comparison helpers to skip until native scan is supported.
-  void compare_gpu_vs_cpu(const std::string& /*query*/,
-                          std::optional<float> /*float_tolerance*/ = std::nullopt)
-  {
-    WARN("duckdb-native tpch scan skipped — legacy duckdb_scan path removed");
-  }
-  bool compare_gpu_vs_cpu_for(int /*num_gpus*/,
-                              const std::string& /*query*/,
-                              std::optional<float> /*float_tolerance*/ = std::nullopt)
-  {
-    WARN("duckdb-native tpch scan skipped — legacy duckdb_scan path removed");
-    return false;  // RUN_TPCH_MGPU returns out of the test on false
-  }
+  // // Disabled: these tests scan native (DuckDB-storage) tables via seq_scan. The
+  // // legacy duckdb_scan path was removed and GPU native scan has no IO backend
+  // // wired in this harness ("missing io_ctx, io_obj"), so every query throws
+  // // "Unsupported scan function: seq_scan" and poisons the shared integration
+  // // DB. Shadow the comparison helpers to skip until native scan is supported.
+  // void compare_gpu_vs_cpu(const std::string& /*query*/,
+  //                         std::optional<float> /*float_tolerance*/ = std::nullopt)
+  // {
+  //   WARN("duckdb-native tpch scan skipped — legacy duckdb_scan path removed");
+  // }
+
+  // bool compare_gpu_vs_cpu_for(int /*num_gpus*/,
+  //                             const std::string& /*query*/,
+  //                             std::optional<float> /*float_tolerance*/ = std::nullopt)
+  // {
+  //   WARN("duckdb-native tpch scan skipped — legacy duckdb_scan path removed");
+  //   return false;  // RUN_TPCH_MGPU returns out of the test on false
+  // }
 };
 
 /**
